Making changes to the kd_tree distance function.
Add a new paramter to the signature of the kd_tree distance function. The
void* would be used to pass the prrts_system_t.space_config object.
prrts_system_t would store the pointer to the PRRTstar object, which would
be passed to the static function PRRTstar::dist_func and inturn used to call
the member distanceFunction. Crazy way to handle C callback from C++.